MOFI/Backend/Producer/Mofi-main/requirements.txt
fastapi[standard]
pydantic
uvicorn[standard]
motor
pymongo
dnspython
python-dotenv
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.2
email-validator
emails
httpx==0.24.1
fastapi-authlib
alembic
cryptography
pyjwt
dynaconf
cloudinary
python-multipart